Refactor Resolver a bit to improve testability
