+++ src/scala/com/github/johnynek/bazel_deps/BUILD
@@ -29,7 +29,17 @@ scala_repl(name = "repl",
 scala_library(name = "resolver",
               srcs = ["Resolver.scala"],
               deps = [
-                 "//3rdparty/jvm/org/apache/maven:maven_aether_provider",
+                  "//3rdparty/jvm/org/typelevel:cats_core",
+                  ":depsmodel",
+                  ":graph",
+              ],
+              visibility = ["//visibility:public"])
+
+scala_library(name = "aether_resolver",
+              srcs = ["AetherResolver.scala"],
+              deps = [
+                  "//3rdparty/jvm/org/typelevel:cats_core",
+                  "//3rdparty/jvm/org/apache/maven:maven_aether_provider",
                   "//3rdparty/jvm/org/eclipse/aether:aether_api",
                   "//3rdparty/jvm/org/eclipse/aether:aether_spi",
                   "//3rdparty/jvm/org/eclipse/aether:aether_connector_basic",
@@ -38,6 +48,7 @@ scala_library(name = "resolver",
                   "//3rdparty/jvm/org/eclipse/aether:aether_transport_http",
                   ":depsmodel",
                   ":graph",
+                  ":resolver",
                   ],
               visibility = ["//visibility:public"])
 
@@ -86,6 +97,7 @@ scala_library(name = "makedeps",
                  ":graph",
                  ":io",
                  ":normalizer",
+                 ":aether_resolver",
                  ":resolver",
                  ":writer",
              ],
